minimal XetRuntime for wasm
